requirements.txt
streamlit>=1.29.0
pandas>=2.0.3
plotly>=5.17.0
PyPDF2>=3.0.1
python-docx>=1.1.0
tiktoken>=0.5.1
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0
faiss-cpu>=1.7.4
numpy>=1.24.0